reset packet sequence on too many old packets
Why:

If the sequencer in the source resets, the livekit packet buffer
does not accept the packets until the sequence numbering has reached the
same number, which can take up to 65535 packets.

When a RTP stream is being forwarded to livekit, this can sometimes
happen if the source stream gets reset but the webrtc session to livekit
is not.

What:

Keep track of consecutive packets that cannot be inserted to the buffer.
If inserting fails enough times (100 for now) consecutively, reset the
buffer to accept the next packet and continue the sequence from there.